refactor(graphql): add `Input` suffix to created input types
Makes it easier to see which symbol in the code is an input, and which
is the EMF EClass.
